New version

+++ compile.sh
@@ -101,8 +101,10 @@ echo "[*] Limpiando kernel..."
 make mrproper
 
 echo "[*] Creando configuracion..."
-cp ../../bcmrpi_defconfig arch/arm/configs/
-ARCH=arm CROSS_COMPILE=${CCPREFIX} make -j $THREADS bcmrpi_defconfig
+if [ ! -f arch/arm/configs/bcmrpi_defconfig ]; then
+	cp ../../bcmrpi_defconfig arch/arm/configs/
+fi
+cp arch/arm/configs/bcmrpi_defconfig .config
 ARCH=arm CROSS_COMPILE=${CCPREFIX} make -j $THREADS oldconfig
 
 echo "[*] Creando menuconfig..."
